feat(planner): Planner 객체 생성 및 기능 추가

+++ src/main/java/christmas/model/EventResult.java
@@ -35,7 +35,7 @@ public class EventResult {
     }
 
 
-    public long calculateExpectedPaymentAmount(int totalOrderAmount) {
+    public long calculateExpectedPaymentAmount(long totalOrderAmount) {
         return totalOrderAmount - eventResult.stream().skip(1).limit(4).mapToLong(Long::longValue).sum();
     }
